+++ CHANGES.rst
@@ -4,6 +4,10 @@ Latest changes
 In development
 --------------
 
+- Ensure that errors in the task generator given to Parallel's call
+  are raised in the results consumming thread.
+  https://github.com/joblib/joblib/pull/1491
+  
 - Adjust codebase to NumPy 2.0 by changing ``np.NaN`` to ``np.nan``
   and importing ``byte_bounds`` from ``np.lib.array_utils``.
   https://github.com/joblib/joblib/pull/1501