+++ CHANGES.rst
@@ -18,6 +18,7 @@ In development
   https://github.com/joblib/joblib/pull/1463
 
 - Vendor cloudpickle 3.0.0 and drop support for Python 3.7.
+  https://github.com/joblib/joblib/pull/1515
 
 Release 1.3.2 -- 2023/08/08
 ---------------------------